src/requirements.txt
Beaker==1.5.4
beautifulsoup4==4.5.1
blivet==2.1.6
Brlapi==0.6.5
chardet==2.3.0
coverage==4.2
cssselect==0.9.2
cupshelpers==1.0
decorator==4.0.10
distro==1.0.3
django-crispy-forms==1.6.1
file-magic==0.3.0
fros==1.1
html5lib==0.999
humanize==0.5.1
iniparse==0.4
IPy==0.81
javapackages==4.7.0
langtable==0.0.36
lxml==3.6.4
Mako==1.0.3
MarkupSafe==0.23
ntplib==0.3.3
ordered-set==2.0.0
pid==2.0.1
Pillow==3.4.2
pwquality==1.3.0
pycups==1.9.72
pycurl==7.43.0
pygobject==3.22.0
pygpgme==0.3
pyinotify==0.9.6
PyIscsi==1.0
pykickstart==2.32
pyparted==3.10.7
PySocks==1.5.6
python-augeas==0.5.0
python-dmidecode==3.12.2
python-meh==0.43
pytz==2016.6.1
pyudev==0.21.0
pyxdg==0.25
requests==2.10.0
requests-file==1.4
requests-ftp==0.3.1
rpm-python==4.13.0rc1
seobject==0.1
sepolicy==1.1
setroubleshoot==1.1
six==1.10.0
slip==0.6.4
slip.dbus==0.6.4
sosreport==3.2.0a1
SSSDConfig==1.14.2
systemd-python==232
urllib3==1.15.1
virtualenv==15.1.0
wrapt==1.10.8
